Extract esc() + fmtKg/fmtQty into shared modules; drop four duplicates
Four near-identical implementations of the HTML-escape function lived in
beamweb.mjs (as escapeHtml), database.mjs (escapeHtml), project-tab.mjs
(esc), and footings-slabs-tab.mjs (esc). Same five-entity map, minor
drift in null-handling but identical observable output. Consolidated
into js/shared/html-utils.mjs as the canonical `esc`. Shell modules that
preferred the verbose name import with an alias (`import { esc as
escapeHtml }`) so call sites need zero changes.

fmtKg and fmtQty display formatters lived only in footings-slabs-tab.mjs
but every Phase 4 assembly tab (Windows, Ceilings, Structural Elements,
... Garage) will need the same rounding contract. Lifted to js/beam/
shared/formatters.mjs before the copy-11x happens. Integer kgCO2e for
emissions, 2dp for quantities, cold-start zeros render as "" not "0.00"
so picker rows stay visually blank until the user engages a dimension.

js/shared/html-utils.mjs is app-agnostic — sits alongside state-manager
and file-handler. js/beam/shared/formatters.mjs is BEAMweb-specific —
lives under js/beam/ because the rounding contract is a BEAM-parity
decision (matches the gSheet display), not a universal one. Pattern for
future extractions: shared concepts go in js/shared/, BEAMweb-specific
helpers that multiple tabs will consume go in js/beam/shared/.

Net: -39 lines across four callers, +45 lines across two new modules.
Next helper planned for extraction: renderCollapsibleSection when the
first Phase 4 assembly tab makes the abstraction concrete. Premature
today since F&S renderGroup and PROJECT renderCollapsibleSection differ
on subtotal + inline config handling.
